+++ src/common/socket.h
@@ -104,8 +104,8 @@ int make_listen_bind(uint32 ip, uint16 port);
 int make_connection(uint32 ip, uint16 port);
 int realloc_fifo(int fd, unsigned int rfifo_size, unsigned int wfifo_size);
 int realloc_writefifo(int fd, size_t addition);
-int WFIFOSET(int fd, int len);
-int RFIFOSKIP(int fd, int len);
+int WFIFOSET(int fd, size_t len);
+int RFIFOSKIP(int fd, size_t len);
 
 int do_sendrecv(int next);
 int do_parsepacket(void);
